= Moved variables and types to their own file
